+++ src/modules/general-affair/entities/dept-catering.vw.ts
@@ -1,0 +1,18 @@
+import { Field, Int, ObjectType } from 'type-graphql';
+import { BaseEntity, Entity, PrimaryColumn } from 'typeorm';
+
+@Entity('ANG_DEPT_CATERING_V')
+@ObjectType()
+export class DeptCateringView extends BaseEntity {
+  @Field(() => Int)
+  @PrimaryColumn({ name: 'ID' })
+  id!: number;
+
+  @Field()
+  @PrimaryColumn({ name: 'DEPT_CODE' })
+  deptCode!: string;
+
+  @Field()
+  @PrimaryColumn({ name: 'DEPT_NAME' })
+  deptName!: string;
+}