+++ CHANGELOG.md
@@ -1,11 +1,19 @@
-<a name="2.37.0"></a>
+<a name="2.38.0"></a>
 
-# [2.37.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/232) (2022-07-27)
+# [2.38.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/238) (2022-07-29)
 
 ### Features
 
-- [**M019**] **gis**: add column lot batch no for gis header.
+- [**M019**] **gis**: add column notes and lot batch no for gis header.
 - [**M019**] **gis**: add column shift for gis operator.
+
+<a name="2.37.0"></a>
+
+# [2.37.0](https://gitlab.com/atjdev/ezio-api/-/merge_requests/236) (2022-07-28)
+
+### Features
+
+- [**G001**] **bon-makan**: add dept catering, add column adjustment makan siang dan sore.
 
 <a name="2.36.0"></a>